Handle writability in meta stores
