+++ terracotta/drivers/terracotta_driver.py
@@ -5,8 +5,7 @@ The driver to interact with.
 
 import contextlib
 from collections import OrderedDict
-import functools
-from typing import (Any, Callable, Collection, Dict, List, Mapping,
+from typing import (Any, Collection, Dict, List, Mapping,
                     Optional, Sequence, Tuple, TypeVar, Union)
 
 import terracotta
@@ -18,17 +17,6 @@ from terracotta.drivers.base_classes import (KeysType, MetaStore,
 ExtendedKeysType = Union[Sequence[str], Mapping[str, str]]
 ExtendedMultiValueKeysType = Union[Sequence[str], Mapping[str, Union[str, List[str]]]]
 T = TypeVar('T')
-
-
-def requires_writable(fun: Callable[..., T]) -> Callable[..., T]:
-    @functools.wraps(fun)
-    def inner(self: "TerracottaDriver", *args: Any, **kwargs: Any) -> T:
-        if self.meta_store.WRITABLE:
-            return fun(self, *args, **kwargs)
-        else:
-            raise exceptions.DatabaseNotWritable("Database not writable")
-
-    return inner
 
 
 def squeeze(iterable: Collection[T]) -> T:
@@ -70,7 +58,6 @@ class TerracottaDriver:
         """
         return self.meta_store.key_names
 
-    @requires_writable
     def create(self, keys: Sequence[str], *,
                key_descriptions: Mapping[str, str] = None) -> None:
         """Create a new, empty metadata store.
@@ -183,16 +170,19 @@ class TerracottaDriver:
             path = squeeze(dataset.values())
             metadata = self.compute_metadata(path, max_shape=self.LAZY_LOADING_MAX_SHAPE)
 
-            if self.meta_store.WRITABLE:
+            try:
                 self.insert(keys, path, metadata=metadata)
 
                 # ensure standardized/consistent output (types and floating point precision)
                 metadata = self.meta_store.get_metadata(keys)
                 assert metadata is not None
+            except exceptions.DatabaseNotWritable as exc:
+                raise exceptions.DatabaseNotWritable(
+                    "Lazy loading requires a writable database"
+                ) from exc
 
         return metadata
 
-    @requires_writable
     @requires_connection
     def insert(
         self, keys: ExtendedKeysType,
@@ -227,7 +217,6 @@ class TerracottaDriver:
             metadata=metadata
         )
 
-    @requires_writable
     @requires_connection
     def delete(self, keys: ExtendedKeysType) -> None:
         """Remove a dataset from the meta store.
